Clere class

This class contains all the input parameters to run CLERE.

# S4 method for Clere
show(object)

# S4 method for Clere,ANY,ANY,ANY
[(x, i, j, drop)

# S4 method for sClere
show(object)

# S4 method for sClere,ANY,ANY,ANY
[(x, i, j, drop)

# S4 method for sClere,ANY,ANY,ANY
[(x, i, j) <- value

Details

y

[numeric]: The vector of observed responses.

x

[matrix]: The matrix of predictors.

n

[integer]: The sample size or the number of rows in matrix x.

p

[integer]: The number of variables of the number of columns in matrix x.

g

[integer]: The number or the maximum number of groups considered. Maximum number of groups stands when model selection is required.

nItMC

[numeric]: Number of Gibbs iterations to generate the partitions.

nItEM

[numeric]: Number of SEM/MCEM iterations.

nBurn

[numeric]: Number of SEM iterations discarded before calculating the MLE which is averaged over SEM draws.

dp

[numeric]: Number of iterations between sampled partitions when calculating the likelihood at the end of the run.

nsamp

[numeric]: Number of sampled partitions for calculating the likelihood at the end of the run.

sparse

[logical]: Should a 0 class be imposed to the model?

analysis

[character]: Which analysis is to be performed. Values are "fit", "bic", "aic" and "icl".

algorithm

[character]: The algorithmto be chosen to fit the model. Either the SEM-Gibbs algorithm or the MCEM algorithm. The most efficient algorithm being the SEM-Gibbs approach. MCEM is not available for binary response.

initialized

[logical]: Is set to TRUE when an initial partition and an initial vector of parameters is given by the user.

maxit

[numeric]: An EM algorithm is used inside the SEM to maximize the complete log-likelihood p(y,Z|theta). maxit stands as the maximum number of EM iterations for the internal EM.

tol

[numeric]: Maximum increased in complete log-likelihood for the internal EM (stopping criterion).

seed

[integer]: An integer given as a seed for random number generation. If set to NULL, then a random seed is generated between 1 and 1000.

b

[numeric]: Vector of parameter b. Its size equals the number of group(s).

pi

[numeric]: Vector of parameter pi. Its size equals the number of group(s).

sigma2

[numeric]: Parameter sigma^2.

gamma2

[numeric]: Parameter gamma^2.

itemintercept[numeric]: Parameter beta_0 (intercept).
likelihood

[numeric]: Approximated log-likelihood.

entropy

[numeric]: Approximated entropy.

P

[matrix]: A p x g matrix of posterior probability of membership to the groups. P = E[Z|theta].

theta

[matrix]: A nItEM x (2g+4) matrix containing values of the model parameters and complete data likelihood at each iteration of the SEM/MCEM algorithm

Bw

[matrix]: A p x nsamp matrix which columns are samples from the posterior distribution of Beta (regression coefficients) given the data and the maximum likelihood estimates.

Zw

[matrix]: A p x nsamp matrix which columns are samples from the posterior distribution of Z (groups membership indicators) given the data and the maximum likelihood estimates.

theta0

[numeric]: A 2g+3 length vector containing initial guess of the model parameters. See example for function fitClere.

Z0

[numeric]: A p x 1 vector of integers taking values between 1 and p (number of variables).

Methods

object["slotName"]:

Get the value of the field slotName.

object["slotName"]<-value:

Set value to the field slotName.

plot(x, ...):

Graphical summary for MCEM/SEM-Gibbs estimation.

clusters(object, threshold = NULL, ...):

Returns the estimated clustering of variables.

predict(object, newx, ...):

Returns prediction using a fitted model and a new matrix of design.

summary(object, ...):

summarizes the output of function fitClere.

See also