Pacs class

This class contains all the input parameters to run CLERE.

Details

Y

[numeric]: The vector of observed responses - size n.

X

[matrix]: The matrix of predictors - size n rows and p columns.

lambda

[numeric]: A non-negative penalty term that controls simultaneouly clusetering and sparsity.

betaInput

[numeric]: A vector of initial guess of the model parameters. The authors suggest to use coefficients obtained after fitting a ridge regression with the shrinkage parameter selected using AIC criterion.

epsPACS

[numeric]: A tolerance threshold that control the convergence of the algroithm. The default value fixed in Bondell's initial script is 1e-5.

nItMax

[integer]: Maximum number of iterations in the algorithm.

a0

[numeric]: Fitted intercept.

K

[integer]: Model dimensionality.

Methods

object["slotName"]:

Get the value of the field slotName.

object["slotName"]<-value:

Set value to the field slotName.

See also