Performances SEM algorithm versus MCEM

This data contains four matrices corresponding to four performance indictors used to compare SEM algorithm and three versions of the MCEM algorithm (MCEMA: with 5 MC interations; MCEMB: with 25 MC iterations and MCEMC: 125 MC iterations) as described in the package vignette. The first matrix Pred contains prediction errors; matrix Bias contains the bias over all model parameters, matrix Time contains execution times for the four methods and matrix Liks the log-likelihood reached by each method. These data were used to generate the Table 1. in the package vignette. For more details, please refer to the package vignette. The R script used to create this dataset is clere/inst/doc/SEM_vs_MCEM_simulations.R.

algoComp

Format

A list containing four 200 x 4/5 matrices.

See also